requirements-streamlit.txt
streamlit==1.41.1
numpy==1.26.4
pandas==2.3.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
opencv-python-headless==4.10.0.84
Pillow==11.1.0
joblib==1.5.3
altair==5.5.0
